requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
click==8.2.1
falcon==4.1.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
